+++ src/main/java/com/ing/request/LoanApproverStatus.java
@@ -8,12 +8,12 @@ import lombok.ToString;
 @NoArgsConstructor @Setter @Getter @ToString
 public class LoanApproverStatus {
 
-	private Long id;
+	private Integer id;
 	private String role;
 	private String status;
 	private String reason;
 	
-	public LoanApproverStatus(Long id, String role, String status, String reason) {
+	public LoanApproverStatus(Integer id, String role, String status, String reason) {
 		super();
 		this.id = id;
 		this.role = role;
